implemented checkTestCaseId, checkAttachmentId

+++ api/src/main/java/com/testquack/api/resources/TestCaseResource.java
@@ -1,5 +1,6 @@
 package com.testquack.api.resources;
 
+import com.testquack.api.utils.APIValidation;
 import com.testquack.beans.Attachment;
 import com.testquack.beans.Filter;
 import com.testquack.beans.Issue;
@@ -61,6 +62,21 @@ public class TestCaseResource extends BaseCrudResource<TestCase> {
     public TestCaseTree findFilteredTree(@ApiParam(value = "Project Id", required = true) @PathParam("projectId") String projectId) {
 System.out.println("TestCaseResource::findFiltered - projectId: " + projectId);
 System.out.flush();
+
+        if (APIValidation.checkProjectId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            projectId) == false) {
+
+            System.out.println("TestCaseResource::findFilteredTree: checkProject returned FALSE - did NOT find project id");
+            System.out.flush();
+
+
+            TestCaseTree tcTree = null;
+            return tcTree;
+        }
+
         return service.findFilteredTree(getUserSession(), projectId, (TestcaseFilter) initFilter(request));
     }
 
@@ -75,6 +91,33 @@ System.out.flush();
 System.out.println("TestCaseResource::upload - testcaseId: " + testcaseId + ", fileDetail: " + fileDetail);
 System.out.println("TestCaseResource::upload - uploadedInputStream: " + uploadedInputStream);
 System.out.flush();
+
+        if (APIValidation.checkProjectId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            projectId) == false) {
+
+            System.out.println("TestCaseResource::upload: checkTestCase returned FALSE - did NOT find project id");
+            System.out.flush();
+
+            TestCase tc = null;
+            return tc;
+        }
+        if (APIValidation.checkTestCaseId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            testcaseId) == false) {
+
+            System.out.println("TestCaseResource::upload: checkTestCase returned FALSE - did NOT find testcase id");
+            System.out.flush();
+
+            TestCase tc = null;
+            return tc;
+        }
+
+
         if (fileDetail == null) {
             throw new EntityValidationException();
         }
@@ -90,6 +133,43 @@ System.out.flush();
             @PathParam("attachmentId") final String attachmentId) {
 System.out.println("TestCaseResource::downloadAttachment - testcaseId: " + testcaseId + ", attachmentId: " + attachmentId);
 System.out.flush();
+
+        if (APIValidation.checkProjectId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            projectId) == false) {
+
+            System.out.println("TestCaseResource::downloadAttachment: checkTestCase returned FALSE - did NOT find project id");
+            System.out.flush();
+
+            Response response = null;
+            return response;
+        }
+        if (APIValidation.checkTestCaseId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            testcaseId) == false) {
+
+            System.out.println("TestCaseResource::downloadAttachment: checkTestCase returned FALSE - did NOT find testcase id");
+            System.out.flush();
+
+            Response response = null;
+            return response;
+        }
+        if (APIValidation.checkAttachmentId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            attachmentId) == false) {
+
+            System.out.println("TestCaseResource::downloadAttachment: checkTestCase returned FALSE - did NOT find attachment id");
+            System.out.flush();
+
+            Response response = null;
+            return response;
+        }
 
         Attachment attachment = service.getAttachment(getUserSession(), projectId, testcaseId, attachmentId);
         try {
@@ -110,6 +190,44 @@ System.out.flush();
             @PathParam("attachmentId") final String attachmentId) throws IOException {
 System.out.println("TestCaseResource::deleteAttachment - testcaseId: " + testcaseId + ", attachmentId: " + attachmentId);
 System.out.flush();
+
+        if (APIValidation.checkProjectId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            projectId) == false) {
+
+            System.out.println("TestCaseResource::deleteAttachment: checkTestCase returned FALSE - did NOT find project id");
+            System.out.flush();
+
+            TestCase tc = null;
+            return tc;
+        }
+        if (APIValidation.checkTestCaseId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            testcaseId) == false) {
+
+            System.out.println("TestCaseResource::deleteAttachment: checkTestCase returned FALSE - did NOT find testcase id");
+            System.out.flush();
+
+            TestCase tc = null;
+            return tc;
+        }
+        if (APIValidation.checkAttachmentId(getService().getMongoReplicaSet(),
+            getService().getMongoUsername(),
+            getService().getMongoPassword(),
+            getService().getMongoDBName(),
+            attachmentId) == false) {
+
+            System.out.println("TestCaseResource::deleteAttachment: checkTestCase returned FALSE - did NOT find attachment id");
+            System.out.flush();
+
+            TestCase tc = null;
+            return tc;
+        }
+
         return service.deleteAttachment(getUserSession(), projectId, testcaseId, attachmentId);
     }
